+++ google2/src/main/scala/org/broadinstitute/dsde/workbench/google2/KubernetesService.scala
@@ -4,7 +4,7 @@ import java.nio.file.Path
 
 import cats.effect.concurrent.Semaphore
 import cats.effect.{Async, Blocker, ContextShift, Effect, Resource, Timer}
-import cats.mtl.ApplicativeAsk
+import cats.mtl.Ask
 import com.google.api.services.container.ContainerScopes
 import io.chrisdavenport.log4cats.StructuredLogger
 import org.broadinstitute.dsde.workbench.google2.GKEModels.KubernetesClusterId
@@ -20,11 +20,11 @@ trait KubernetesService[F[_]] {
 
   // namespaces group resources, and allow our list/get/update API calls to be segmented. This can be used on a per-user basis, for example
   def createNamespace(clusterId: KubernetesClusterId, namespace: KubernetesNamespace)(
-    implicit ev: ApplicativeAsk[F, TraceId]
+    implicit ev: Ask[F, TraceId]
   ): F[Unit]
 
   def deleteNamespace(clusterId: KubernetesClusterId, namespace: KubernetesNamespace)(
-    implicit ev: ApplicativeAsk[F, TraceId]
+    implicit ev: Ask[F, TraceId]
   ): F[Unit]
 
   // A Kubernetes service account is an automatically enabled authenticator that uses signed bearer tokens to verify requests.
@@ -32,39 +32,39 @@ trait KubernetesService[F[_]] {
   def createServiceAccount(clusterId: KubernetesClusterId,
                            serviceAccount: KubernetesServiceAccount,
                            namespaceName: KubernetesNamespace)(
-    implicit ev: ApplicativeAsk[F, TraceId]
+    implicit ev: Ask[F, TraceId]
   ): F[Unit]
 
   // pods represent a set of containers
   def createPod(clusterId: KubernetesClusterId, pod: KubernetesPod, namespace: KubernetesNamespace)(
-    implicit ev: ApplicativeAsk[F, TraceId]
+    implicit ev: Ask[F, TraceId]
   ): F[Unit]
 
   def listPodStatus(clusterId: KubernetesClusterId, namespace: KubernetesNamespace)(
-    implicit ev: ApplicativeAsk[F, TraceId]
+    implicit ev: Ask[F, TraceId]
   ): F[List[KubernetesPodStatus]]
 
   // certain services allow us to expose various containers via a matching selector
   def createService(clusterId: KubernetesClusterId, service: KubernetesServiceKind, namespace: KubernetesNamespace)(
-    implicit ev: ApplicativeAsk[F, TraceId]
+    implicit ev: Ask[F, TraceId]
   ): F[Unit]
 
   def getServiceExternalIp(clusterId: KubernetesClusterId, namespace: KubernetesNamespace, serviceName: ServiceName)(
-    implicit ev: ApplicativeAsk[F, TraceId]
+    implicit ev: Ask[F, TraceId]
   ): F[Option[IP]]
 
   def createRole(clusterId: KubernetesClusterId, role: KubernetesRole, namespace: KubernetesNamespace)(
-    implicit ev: ApplicativeAsk[F, TraceId]
+    implicit ev: Ask[F, TraceId]
   ): F[Unit]
 
   def createRoleBinding(clusterId: KubernetesClusterId,
                         roleBinding: KubernetesRoleBinding,
                         namespace: KubernetesNamespace)(
-    implicit ev: ApplicativeAsk[F, TraceId]
+    implicit ev: Ask[F, TraceId]
   ): F[Unit]
 
   def createSecret(clusterId: KubernetesClusterId, namespace: KubernetesNamespace, secret: KubernetesSecret)(
-    implicit ev: ApplicativeAsk[F, TraceId]
+    implicit ev: Ask[F, TraceId]
   ): F[Unit]
 }